fix: read the Windows-specific optional header fields
The optional header reader filled in the standard fields and the data
directories but none of the Windows-specific fields between them, so
ImageBase, SizeOfImage, Subsystem, DllCharacteristics and the rest were always
0. 1.2.5 read them into COFF_Header, and they were lost when that type was
folded into OptionalHeader. BaseOfData was also read for PE32+, which has no
such field, picking up the low half of ImageBase, and left at 0 for PE32,
which has one.

Every field is now read at the offset the PE format gives for PE32 and PE32+.
ImageBase and the stack and heap sizes are 64 bits wide in PE32+, so those
properties become ulong, as they were on COFF_Header.

Icon and cursor reading only uses the resource directory, so it is unaffected.
The tests pin each field at its offset and width for both formats, and compare
against System.Reflection.Metadata on the PE32 fixture and on the runtime's
core library, which is PE32+. user32.dll now reports an ImageBase of
0x180000000 instead of 0.

1098 tests pass.

+++ Ico.Reader.PeDecoder/Models/OptionalHeader.cs
@@ -50,14 +50,14 @@ public sealed class OptionalHeader
     public uint BaseOfCode { get; init; }
 
     /// <summary>
-    /// Address of the first data section relative to the image base. Present on PE32 only.
+    /// Address of the first data section relative to the image base. Present on PE32 only, and 0 for PE32+.
     /// </summary>
     public uint BaseOfData { get; init; }
 
     /// <summary>
     /// Preferred address of the first byte of the image once loaded.
     /// </summary>
-    public uint ImageBase { get; init; }
+    public ulong ImageBase { get; init; }
 
     /// <summary>
     /// Alignment in bytes of sections once loaded.
@@ -132,22 +132,22 @@ public sealed class OptionalHeader
     /// <summary>
     /// Stack size to reserve.
     /// </summary>
-    public uint SizeOfStackReserve { get; init; }
+    public ulong SizeOfStackReserve { get; init; }
 
     /// <summary>
     /// Stack size to commit up front.
     /// </summary>
-    public uint SizeOfStackCommit { get; init; }
+    public ulong SizeOfStackCommit { get; init; }
 
     /// <summary>
     /// Local heap size to reserve.
     /// </summary>
-    public uint SizeOfHeapReserve { get; init; }
+    public ulong SizeOfHeapReserve { get; init; }
 
     /// <summary>
     /// Local heap size to commit up front.
     /// </summary>
-    public uint SizeOfHeapCommit { get; init; }
+    public ulong SizeOfHeapCommit { get; init; }
 
     /// <summary>
     /// Reserved and required to be zero.
